+++ pkgs/gcloud/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.9.0
+
+- Support `orderingKey` on pub/sub's `Message` type.
+- Support the latest version `^15.0.0` of the `googleapis` package.
+
 ## 0.8.19
 
 - Support the latest `package:googleapis`.